feat(trade): refactor input handling by integrating InputNumberControl and removing OrderFormInput
