+++ src/features/trade/components/OrderForm/TrailingQuote.tsx
@@ -1,0 +1,13 @@
+import { useShallowInstrumentStore } from "@/store/trade/instrument";
+
+const TrailingQuote = () => {
+  const quote = useShallowInstrumentStore((s) => s.assetMeta?.quote);
+
+  return (
+    <div className="flex items-center gap-x-2">
+      <span className="text-neutral-300 text-sm font-medium">{quote}</span>
+    </div>
+  );
+};
+
+export default TrailingQuote;
